+++ reactfront/src/components/DeleteAccount.js
@@ -9,32 +9,33 @@ const DeleteAccount = ({ accessToken, userID, navigate, onAccountDeleted }) => {
 
   const handleDeleteAccount = async () => {
     setLoading(true);
-
-    try {
-      const response = await axios.delete(`https://demolaravel.ddns.net/api/user/${userID}`, {
-        headers: {
-          Authorization: `Bearer ${accessToken}`
+    const confirmed = window.confirm("¿Está seguro de que desea eliminar esta cuenta?");
+    if (confirmed) {
+      try {
+        const response = await axios.delete(`https://demolaravel.ddns.net/api/user/${userID}`, {
+          headers: {
+            Authorization: `Bearer ${accessToken}`
+          }
+        });
+        
+        console.log('Delete account response:', response.data);
+        /* Función para restablecer a null los props, cuando borremos una cuenta, 
+        si lo hacemos desde el panel de administrador no enviamos este parámetro*/
+        if (onAccountDeleted){
+          onAccountDeleted();
         }
-      });
-      
-      console.log('Delete account response:', response.data);
-      /* Función para restablecer a null los props, cuando borremos una cuenta, 
-      si lo hacemos desde el panel de administrador no enviamos este parámetro*/
-      if (onAccountDeleted){
-        onAccountDeleted();
+        alert('¡Cuenta eliminada con éxito!');
+        if (navigate){
+          navigate('/');
+        }else{
+          defaultNavigate('/admin');
+        }
+      } catch (error) {
+        console.error('Error al eliminar la cuenta:', error);
+        alert('Se ha producido un error al eliminar la cuenta. Por favor, intenta nuevamente más tarde.');
       }
-      alert('¡Cuenta eliminada con éxito!');
-      if (navigate){
-        navigate('/');
-      }else{
-        defaultNavigate('/admin');
-      }
-    } catch (error) {
-      console.error('Error al eliminar la cuenta:', error);
-      alert('Se ha producido un error al eliminar la cuenta. Por favor, intenta nuevamente más tarde.');
     }
-
-    setLoading(false);
+      setLoading(false);
   };
   
   return (
